+++ src/hooks/useSemiPersistentState.js
@@ -1,0 +1,16 @@
+import { useState, useEffect } from 'react'
+
+const useSemiPersistentState = () => {
+
+  const [todoList, setTodoList] = useState(
+    JSON.parse(localStorage.getItem('savedTodoList')) || []
+  )
+
+  useEffect(() => {
+    localStorage.setItem('savedTodoList', JSON.stringify(todoList))
+  }, [todoList])
+
+  return [todoList, setTodoList]
+}
+
+export default useSemiPersistentState
